+++ VpnHood.Client.App.UI.Html/_publish.ps1
@@ -6,7 +6,7 @@ $dest1= "$solutionDir\VpnHood.Client.App.Android\Assets\SPA.zip";
 $dest2= "$solutionDir\VpnHood.Client.App.Win\Resources\SPA.zip";
 
 # build output
-& npm run build --prefix $PSScriptRoot
+./_publish.bat
 if ($LASTEXITCODE -gt 0) { Write-Host ("Error code: " + $lastexitcode) -ForegroundColor Red; pause;}
 
 # zip